Separate DeepestNotMeContactTestResultCallback
